+++ nesferratu-core/src/cpu/instructions/addressing.rs
@@ -1,26 +1,26 @@
 use crate::{BusMessage, cpu::{AddrDelegateReturn, CpuState, Operand}};
 
-pub fn acc(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn acc(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     AddrDelegateReturn::Return(Operand::Implied)
 }
 
-pub fn imm(s: &mut CpuState, _cycle: usize) -> AddrDelegateReturn {
+pub fn imm(s: &mut CpuState, _cycle: u8) -> AddrDelegateReturn {
     AddrDelegateReturn::Return(Operand::Immediate(s.o1))
 }
 
-pub fn abs(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn abs(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     s.addr = s.o1 as u16;
     s.addr |= (s.o2 as u16) << 8;
 
     AddrDelegateReturn::Return(Operand::Address(s.addr))
 }
 
-pub fn zp(s: &mut CpuState, _cycle: usize) -> AddrDelegateReturn {
+pub fn zp(s: &mut CpuState, _cycle: u8) -> AddrDelegateReturn {
     s.addr = s.o1 as u16;
     AddrDelegateReturn::Return(Operand::Address(s.addr))
 }
 
-pub fn zp_x(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn zp_x(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => AddrDelegateReturn::Yield(BusMessage::Nop),
         2 => {
@@ -31,7 +31,7 @@ pub fn zp_x(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     }
 }
 
-pub fn zp_y(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn zp_y(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => AddrDelegateReturn::Yield(BusMessage::Nop),
         2 => {
@@ -42,7 +42,7 @@ pub fn zp_y(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     }
 }
 
-pub fn abs_x(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn abs_x(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     s.addr = s.o1 as u16;
     s.addr |= (s.o2 as u16) << 8;
     s.addr += s.regs.x as u16;
@@ -50,7 +50,7 @@ pub fn abs_x(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     AddrDelegateReturn::Return(Operand::Address(s.addr))
 }
 
-pub fn abs_y(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn abs_y(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     s.addr = s.o1 as u16;
     s.addr |= (s.o2 as u16) << 8;
     s.addr += s.regs.y as u16;
@@ -58,11 +58,11 @@ pub fn abs_y(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     AddrDelegateReturn::Return(Operand::Address(s.addr))
 }
 
-pub fn imp(_s: &mut CpuState, _cycle: usize) -> AddrDelegateReturn {
+pub fn imp(_s: &mut CpuState, _cycle: u8) -> AddrDelegateReturn {
     AddrDelegateReturn::Return(Operand::Implied)
 }
 
-pub fn rel(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn rel(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     let offset = s.o1 as i8;
     let mut new_pc = s.regs.pc as i32;
     new_pc += offset as i32;
@@ -70,7 +70,7 @@ pub fn rel(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     AddrDelegateReturn::Return(Operand::Address(new_pc as u16))
 }
 
-pub fn ind_x(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn ind_x(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => {
             // fetch zero page address
@@ -96,7 +96,7 @@ pub fn ind_x(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     }
 }
 
-pub fn ind_y(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn ind_y(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => {
             // fetch LSB of address from zero page
@@ -123,7 +123,7 @@ pub fn ind_y(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     }
 }
 
-pub fn ind(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn ind(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => {
             // fetch new PC LSB at operand
@@ -151,7 +151,7 @@ pub fn ind(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     }
 }
 
-pub fn abs_x_extra(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn abs_x_extra(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => {
             s.addr = s.o1 as u16;
@@ -173,7 +173,7 @@ pub fn abs_x_extra(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     }
 }
 
-pub fn abs_y_extra(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn abs_y_extra(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => {
             s.addr = s.o1 as u16;
@@ -195,7 +195,7 @@ pub fn abs_y_extra(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
     }
 }
 
-pub fn ind_y_extra(s: &mut CpuState, cycle: usize) -> AddrDelegateReturn {
+pub fn ind_y_extra(s: &mut CpuState, cycle: u8) -> AddrDelegateReturn {
     match cycle {
         1 => {
             // fetch LSB of address from zero page
